Different fixes for new protos.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1055 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 4d6eb0b9cfed4cbb2766865c542af778b9a342c9

+++ Source/Critter.h
@@ -112,7 +112,7 @@ public:
     CLASS_PROPERTY( ScriptString *, PassHash );
     CLASS_PROPERTY( uint, CrType );
     CLASS_PROPERTY( uint, CrTypeAlias );
-    CLASS_PROPERTY( uchar, Cond );
+    CLASS_PROPERTY( int, Cond ); // enum CritterCondition
     CLASS_PROPERTY( bool, ClientToDelete );
     CLASS_PROPERTY( uint, MultihexBase );
     CLASS_PROPERTY( ushort, WorldX );
